+++ test/default-types.ts
@@ -4,7 +4,7 @@ import { assertSuccess, assertFailure } from './helpers'
 
 describe('Dictionary', () => {
   it('should succeed validating a valid value', () => {
-    const T = t.Dictionary
+    const T = t.UnknownRecord
     assertSuccess(T.decode({}))
     assertSuccess(T.decode([]))
     assertSuccess(T.decode([1]))
@@ -13,29 +13,29 @@ describe('Dictionary', () => {
   })
 
   it('should fail validating an invalid value', () => {
-    const T = t.Dictionary
-    assertFailure(T.decode('s'), ['Invalid value "s" supplied to : Dictionary'])
-    assertFailure(T.decode(1), ['Invalid value 1 supplied to : Dictionary'])
-    assertFailure(T.decode(true), ['Invalid value true supplied to : Dictionary'])
-    assertFailure(T.decode(null), ['Invalid value null supplied to : Dictionary'])
-    assertFailure(T.decode(undefined), ['Invalid value undefined supplied to : Dictionary'])
+    const T = t.UnknownRecord
+    assertFailure(T.decode('s'), ['Invalid value "s" supplied to : UnknownRecord'])
+    assertFailure(T.decode(1), ['Invalid value 1 supplied to : UnknownRecord'])
+    assertFailure(T.decode(true), ['Invalid value true supplied to : UnknownRecord'])
+    assertFailure(T.decode(null), ['Invalid value null supplied to : UnknownRecord'])
+    assertFailure(T.decode(undefined), ['Invalid value undefined supplied to : UnknownRecord'])
   })
 })
 
-describe('Array', () => {
+describe('UnknownArray', () => {
   it('should succeed decoding a valid value', () => {
-    const T = t.Array
+    const T = t.UnknownArray
     assertSuccess(T.decode([]))
     assertSuccess(T.decode([1]))
   })
 
   it('should fail decoding an invalid value', () => {
-    const T = t.Array
-    assertFailure(T.decode('s'), ['Invalid value "s" supplied to : Array'])
-    assertFailure(T.decode(1), ['Invalid value 1 supplied to : Array'])
-    assertFailure(T.decode(true), ['Invalid value true supplied to : Array'])
-    assertFailure(T.decode(null), ['Invalid value null supplied to : Array'])
-    assertFailure(T.decode(undefined), ['Invalid value undefined supplied to : Array'])
+    const T = t.UnknownArray
+    assertFailure(T.decode('s'), ['Invalid value "s" supplied to : UnknownArray'])
+    assertFailure(T.decode(1), ['Invalid value 1 supplied to : UnknownArray'])
+    assertFailure(T.decode(true), ['Invalid value true supplied to : UnknownArray'])
+    assertFailure(T.decode(null), ['Invalid value null supplied to : UnknownArray'])
+    assertFailure(T.decode(undefined), ['Invalid value undefined supplied to : UnknownArray'])
   })
 })